[Refactoring] Extract command factory to new files
