+++ TestShell/TestShell.vcxproj.filters
@@ -87,6 +87,9 @@
     <ClCompile Include="shell_run_script_test.cpp">
       <Filter>테스트 파일</Filter>
     </ClCompile>
+    <ClCompile Include="command_factory.cpp">
+      <Filter>소스 파일</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="ssd_interface.h">
@@ -108,6 +111,9 @@
       <Filter>테스트 파일</Filter>
     </ClInclude>
     <ClInclude Include="logger.h">
+      <Filter>헤더 파일</Filter>
+    </ClInclude>
+    <ClInclude Include="command_factory.h">
       <Filter>헤더 파일</Filter>
     </ClInclude>
   </ItemGroup>
