+++ TestShell/command_factory.h
@@ -1,0 +1,8 @@
+#pragma once
+#include "command.h"
+
+// command Factory
+class CommandFactory {
+public:
+    static std::unique_ptr<Command> createCommand(const std::vector<std::string>& commandVector, SSDInterface* ssd);
+};
